+++ client/keys.c
@@ -818,7 +818,7 @@ Called by the system between frames for both key up and key down events
 Should NOT be called during an interrupt!
 ===================
 */
-void Key_Event (int key, qboolean down, double time)
+void Key_Event (int key, qboolean down, unsigned time)
 {
 	char	*kb;
 	char	cmd[1024];
@@ -948,7 +948,7 @@ void Key_Event (int key, qboolean down, double time)
 		kb = keybindings[key];
 		if (kb && kb[0] == '+')
 		{
-			Com_sprintf (cmd, sizeof(cmd), "-%s %i %f\n", kb+1, key, time);
+			Com_sprintf (cmd, sizeof(cmd), "-%s %i %i\n", kb+1, key, time);
 			Cbuf_AddText (cmd);
 		}
 		if (keyshift[key] != key)
@@ -956,7 +956,7 @@ void Key_Event (int key, qboolean down, double time)
 			kb = keybindings[keyshift[key]];
 			if (kb && kb[0] == '+')
 			{
-				Com_sprintf (cmd, sizeof(cmd), "-%s %i %f\n", kb+1, key, time);
+				Com_sprintf (cmd, sizeof(cmd), "-%s %i %i\n", kb+1, key, time);
 				Cbuf_AddText (cmd);
 			}
 		}
@@ -975,7 +975,7 @@ void Key_Event (int key, qboolean down, double time)
 		{
 			if (kb[0] == '+')
 			{	// button commands add keynum and time as a parm
-				Com_sprintf (cmd, sizeof(cmd), "%s %i %f\n", kb, key, time);
+				Com_sprintf (cmd, sizeof(cmd), "%s %i %i\n", kb, key, time);
 				Cbuf_AddText (cmd);
 			}
 			else
